+++ resources/python_installation.rb
@@ -5,12 +5,13 @@ provides :python_installation
 
 default_action :create
 
-property :version, String, default: '3.7.4'
+property :version, [String, nil], default: nil
 property :download_directory, [String, nil], default: nil
 property :build_directory, [String, nil], default: nil
 property :install_directory, [String, nil], default: nil
 property :openssl_directory, [String, nil], default: nil
 property :sqlite_directory, [String, nil], default: nil
+property :build_shared, [true, false], default: false
 property :owner, String, default: 'root'
 property :group, String, default: 'root'
 
@@ -19,5 +20,5 @@ action :create do
 end
 
 action_class do
-  include PythonInstall::Helper
+  include PythonInstall::Install
 end