+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for Python Install Cookbook
 
+## 1.1.0
+
+* Added helpers for default version and path
+* Created installer library
+
 ## 1.0.0
 
 * Initial release